requirements.txt
exa-py
python-dotenv

